% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_lavels.R
\name{set_lavels}
\alias{set_lavels}
\title{set_lavels: Assign factor labels to multiple variables}
\usage{
set_lavels(df, label_list)
}
\arguments{
\item{df}{A data.frame.}

\item{label_list}{A named list of label vectors. Each name must match a column in df.}
}
\value{
A data.frame with labeled factor variables.
}
\description{
set_lavels: Assign factor labels to multiple variables
}
